patients-app-api-devops/requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
Pillow>=8.2.0,<8.3.0
uwsgi>=2.0.18,<2.1.0
boto3>=1.12.0,<1.13.0
django-storages>=1.9.1,<1.10.0
python-facepp
emoji
